+++ contracts/common/IERC20.sol
@@ -2,6 +2,7 @@
 pragma solidity ^0.8.0;
 
 interface IERC20 {
-    function totalSupply() external view returns (uint);
-    function balanceOf(address account) external view returns (uint);
+    function totalSupply() external view returns (uint256);
+
+    function balanceOf(address account) external view returns (uint256);
 }
